+++ conversations/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 
 class Messages(models.Model):
-    uid = models.IntegerField()
+    uid = models.BigIntegerField()
     responderId = models.IntegerField()
     recieverId = models.IntegerField()
     message = models.TextField()
